Django/requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.15.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
Django==2.1.2
django-cors-headers==2.4.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mysqlclient==1.3.13
Pillow==5.3.0
pycparser==2.19
PyJWT==1.6.4
PyMySQL==0.9.2
python-http-client==3.1.0
pytz==2018.5
requests==2.20.0
sendgrid==5.6.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
